Revert "Adds custom hook for the preloader, updated form styles, adds Donatio…"
