+++ src/Components/Disclaimer/Disclaimer.jsx
@@ -2,11 +2,13 @@ import styles from "./Disclaimer.module.css";
 
 const Disclaimer = () => {
   return (
-    <p className={styles.disclaimer}>
-      Todos los valores publicados en este sitio web son a fines informativos y
-      tienen un carácter orientativo, por lo que deben ser tomados únicamente a
-      modo de referencia.
-    </p>
+    <footer className={styles.disclaimer}>
+      <p>
+        Todos los valores publicados en este sitio web son a fines informativos
+        y tienen un carácter orientativo, por lo que deben ser tomados
+        únicamente a modo de referencia.
+      </p>
+    </footer>
   );
 };
