+++ src/Components/Disclaimer/Disclaimer.module.css
@@ -1,9 +1,39 @@
 .disclaimer {
+  width: 90%;
+  max-width: 750px;
+  padding: 1rem 0 1rem 0;
+  margin-top: 1rem;
   place-self: center;
   font-size: 0.75rem;
   font-style: italic;
-  font-weight: var(--fw-light-bold);
   letter-spacing: 1.2px;
   line-height: 1.3;
-  max-width: 80ch;
+  color: var(--font-color);
+  font-weight: var(--fw-light-bold);
+  border-top: 1px solid var(--form-separator);
+}
+
+@media (min-width: 750px) {
+  .disclaimer {
+    width: 80%;
+    line-height: 1.4;
+  }
+}
+
+@media (min-width: 1000px) {
+  .disclaimer {
+    width: 70%;
+  }
+}
+
+@media (min-width: 1250px) {
+  .disclaimer {
+    width: 60%;
+  }
+}
+
+@media (min-width: 1500px) {
+  .disclaimer {
+    width: 50%;
+  }
 }
