--- src/Components/Donation/Donation.module.css
@@ -1,11 +1,0 @@
-.donation {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 0.5rem;
-  font-size: 0.9rem;
-  font-weight: var(--fw-bold);
-  text-align: center;
-  font-style: normal;
-  color: var(--font-color);
-}
