+++ src/Components/Form/Form.module.css
@@ -78,8 +78,8 @@
   padding: 0.5rem 0.25rem;
   width: 50%;
   min-width: 100px;
-  transition: border-color 0.2s ease, color 0.2s ease,
-    background-color 0.2s ease;
+  transition: border-color 0.3s ease, color 0.3s ease,
+    background-color 0.3s ease;
 
   color: var(--btn-primary-fcolor);
   background-color: var(--btn-primary-bgcolor);
@@ -94,20 +94,21 @@
 
 /* MEDIA QUERIES */
 
-@media (min-width: 600px) {
+@media (min-width: 500px) {
   .shifts {
     display: grid;
-    grid-template-columns: 1fr minmax(65px, 0.5fr) 1fr minmax(65px, 0.5fr);
-    gap: 0.75rem 1.2rem;
+    grid-template-columns: repeat(4, 1fr);
+    gap: 0.5rem;
     align-items: center;
   }
+}
 
+@media (min-width: 500px) {
   .button-container {
     margin-top: 0.5rem;
     width: 80%;
   }
 }
-
 @media (min-width: 750px) {
   .form {
     padding: 1rem 1rem;
@@ -121,12 +122,6 @@
 
   .months-container::after {
     transform: translateY(0.75rem);
-  }
-}
-
-@media (min-width: 850px) {
-  .shifts {
-    gap: 0.75rem 1.5rem;
   }
 }
