+++ src/Components/Input/Input.module.css
@@ -1,12 +1,13 @@
 .label {
   position: relative;
   font-style: italic;
-  font-size: 0.9rem;
+  font-size: 1rem;
   color: var(--font-color);
   font-weight: var(--fw-light-bold);
   display: flex;
   gap: 10px;
   align-items: center;
+  min-width: 155px;
 }
 
 .info {
@@ -27,7 +28,7 @@
   background-color: var(--input-bgcolor);
   width: 100%;
   min-width: 50px;
-  max-width: 70px;
+  max-width: 80px;
   outline: none;
   border: 2px solid var(--input-border-color);
   border-radius: 5px;
@@ -100,12 +101,7 @@
 }
 
 @media (min-width: 750px) {
-  .label {
-    font-size: 1rem;
-  }
-
   .input {
     place-self: center;
-    max-width: 80px;
   }
 }
