+++ src/Components/Months/Months.jsx
@@ -1,5 +1,5 @@
 import styles from "./Months.module.css";
-import { MONTHS } from "../../utils/constants";
+import { MONTHS } from "../../Utils/constants";
 import Month from "./Month";
 
 const Months = ({ onSelectMonth, id }) => {
@@ -9,7 +9,6 @@ const Months = ({ onSelectMonth, id }) => {
       name={id}
       id={id}
       onChange={(e) => onSelectMonth(e)}
-      aria-label="Selecciona un mes"
     >
       <option value="" className={styles.option}>
         Seleccioná un mes
