+++ src/Components/Result/Result.module.css
@@ -45,8 +45,8 @@
   max-width: 160px;
   max-height: 35px;
   padding: 0.15rem 0.25rem;
-  transition: border-color 0.2s ease, color 0.2s ease,
-    background-color 0.2s ease;
+  transition: border-color 0.3s ease, color 0.3s ease,
+    background-color 0.3s ease;
   color: var(--btn-details-fcolor);
   background-color: var(--btn-details-bgcolor);
   border: 2px solid var(--btn-details-border-color);
