+++ src/Components/Result/ResultsDetails.jsx
@@ -1,6 +1,6 @@
 import styles from "./ResultDetails.module.css";
 import Detail from "../Detail/Detail";
-import { formatNumber } from "../../utils/formatNumber";
+import { formatNumber } from "../../Utils/formatNumber";
 
 const ResultDetails = ({ details, classes }) => {
   const curedDetails = Object.entries(details).filter((detail) => {
